Prevent crashes by only loading the next room if there are no more dialogs to dismiss.

+++ sw/firmware/src/agi/screen.h
@@ -14,8 +14,8 @@ public:
     void line(int x0, int y0, int x1, int y1, uint8_t color);
     void fill(int x, int y, uint8_t from_color, uint8_t to_color);
 
-    static constexpr size_t WIDTH = 160;
-    static constexpr size_t HEIGHT = 168;
+    static constexpr int WIDTH = 160;
+    static constexpr int HEIGHT = 168;
     uint8_t buffer[WIDTH*HEIGHT];
 };
